requirements.txt
langchain
langchain-community 
langchain-openai
chromadb
openai
python-dotenv
tiktoken
fastapi
uvicorn